+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Improved performance of CSV and JSON encoding
 - Added support for emitting warnings from historical API response headers, such as for
   future deprecations
+- Added `symbol_map` method to the `Resolution` struct returned by `symbology::resolve`
+  that returns a `TsSymbolMap`
 - Added `PartialEq` and `Eq` implementations for parameter builder classes
 
 ## 0.4.2 - 2023-10-23